redis缓存修改特定key的过期时间

+++ src/main/java/com/example/boot/model/Father.java
@@ -10,5 +10,5 @@ import lombok.Data;
 public class Father {
 
     private String fatherName;
-    private Short fatherAge;
+    private Long fatherAge;
 }
